:recycle: separrate App and Bingo Sheet
